CRUD methods implemented for City entity test modified;
doGet method in CityServlet added;

CityServlet modified;

logging fixed;

addCity, updateCity, deleteCity, getCityById added;
tests added;

CityServlet added;

update City added;

json file added;

City entity added;

JdbcCityDao modified;

+++ src/test/java/com/alexa/worldservice/mapper/CountryMapperTest.java
@@ -37,6 +37,6 @@ public class CountryMapperTest {
         assertEquals("Europe", country.getContinent());
         assertEquals("Eastern Europe", country.getRegion());
         assertEquals(603700.0, country.getSurfaceArea(),0.00);
-        assertEquals(50456000,country.getPopulation());
+        assertEquals(50456000,(int)country.getPopulation());
     }
 }
